requirements.txt
google-genai
faiss-cpu~=1.10.0
beautifulsoup4